usdc-ai-optimiser/requirement.txt
# requirements.txt
aiohttp==3.9.0
asyncio-throttle==1.0.2
anthropic==0.3.11
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.4
numpy==1.24.4
web3==6.11.0
requests==2.31.0

